tools/platform_mcp_server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
psycopg2-binary==2.9.9
pymysql==1.1.0
pinecone>=5.0.0
openai>=1.0.0
weaviate-client>=4.0.0
qdrant-client>=1.7.0
chromadb>=0.4.0
pandas>=2.0.0
boto3>=1.28.0
snowflake-connector-python[pandas]>=3.10.0
google-cloud-bigquery>=3.14.0
google-cloud-storage>=2.14.0
pymssql>=2.2.11
databricks-sql-connector>=3.0.0
slack-sdk>=3.27.0
PyGithub>=2.2.0
notion-client>=2.2.0
azure-storage-blob>=12.19.0
